# backend/api-gateway/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SportBook API",
    "version": "4.0.0",
    "description": "REST API of the SportBook venue booking platform. The API Gateway forwards requests over RabbitMQ RPC to the User, Venue, Booking, Notification and Payment microservices.\n\n**Auth:** call `POST /user/login`, then click **Authorize** and paste the token.\n\n**Payments:** when `STRIPE_SECRET_KEY` is configured, paid slots are booked as `pending_payment` and `POST /bookings/create` returns a `checkout_url` (Stripe Checkout). The slot is held for 30 minutes; the booking becomes `booked` once Stripe confirms the payment (webhook or `GET /payments/verify` on return). Cancelling a paid booking refunds it in full."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Slot is not available"
          }
        }
      },
      "VenueType": {
        "type": "string",
        "enum": [
          "football_field",
          "tennis_court",
          "basketball_court"
        ]
      },
...
      "Booking": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "venue_id": {
            "type": "string",
            "format": "uuid"
          },
          "slot_id": {
            "type": "string",
            "format": "uuid"
          },
          "start_time": {
            "type": "string",
            "format": "date-time"
          },
          "end_time": {
            "type": "string",
            "format": "date-time"
          },
          "price": {
            "type": "string",
            "example": "60.00"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending_payment",
              "booked",
              "cancelled"
            ]
          },
          "venue": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "type": {
                "$ref": "#/components/schemas/VenueType"
              },
              "location": {
                "type": "string"
              }
            }
          },
          "review": {
            "type": "object",
            "nullable": true,
            "properties": {
              "rating": {
                "type": "integer"
              },
              "comment": {
                "type": "string",
                "nullable": true
              }
            }
          },
          "hold_expires_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Until when the slot is held while awaiting payment"
          },
          "paid_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "refunded_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "cancel_reason": {
            "type": "string",
            "nullable": true,
            "enum": [
              "user",
              "payment_expired",
              "payment_failed",
              "late_payment"
            ]
          }
        }
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Missing or invalid token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Not allowed for this user",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
...
  },
  "paths": {
...
    "/bookings/create": {
      "post": {
        "tags": [
          "Bookings"
        ],
        "summary": "Book a slot",
        "description": "The slot is reserved atomically: of two concurrent requests for the same slot only one succeeds, the other gets 409. The price is fixed at booking time. With payments enabled the booking is returned as `pending_payment` together with `checkout_url`; redirect the player there. Without payments (or for admins / free slots) it is `booked` immediately.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "slot_id"
                ],
                "properties": {
                  "slot_id": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Booked, or awaiting payment",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Booking"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "checkout_url": {
                          "type": "string",
                          "nullable": true,
                          "example": "https://checkout.stripe.com/c/pay/cs_test_…"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Slot is not available",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Payment service unavailable (slot released)"
          }
        }
      }
    },
    "/bookings/{id}": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "summary": "Bookings of the current user (userId must match the token)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "User id (GET) — must match the token",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Booking"
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "delete": {
        "tags": [
          "Bookings"
        ],
        "summary": "Cancel an own booking (a paid booking is refunded in full first)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Booking id (DELETE)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "refund": {
                      "type": "object",
                      "nullable": true,
                      "properties": {
                        "amount": {
                          "type": "number"
                        },
                        "status": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Already cancelled"
          },
          "502": {
            "description": "Refund failed — the booking was not cancelled"
          }
        }
      }
    },
...
    "/bookings/{bookingId}/review": {
      "post": {
        "tags": [
          "Reviews",
          "Bookings"
        ],
        "summary": "Rate a played game (creates or updates the review)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "bookingId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "rating"
                ],
                "properties": {
                  "rating": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 5
                  },
                  "comment": {
                    "type": "string",
                    "maxLength": 1000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "The game has not been played yet"
          }
        }
      }
    },
...
    "/bookings/{bookingId}/pay": {
      "post": {
        "tags": [
          "Payments",
          "Bookings"
        ],
        "summary": "New Checkout Session for a booking that awaits payment (\"Pay now\")",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "bookingId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Checkout URL (or `{ booked: true }` when payments are disabled)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "checkout_url": {
                      "type": "string"
                    },
                    "hold_expires_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "booked": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Booking does not await payment"
          },
          "503": {
            "description": "Payment service unavailable"
          }
        }
      }
    },
...
  }
}